# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EgliseHub API",
        "description": "Plateforme web multi-ministères chrétiens — République centrafricaine",
        "contact": {
            "name": "EgliseHub Support",
            "email": "admin@eglisehub.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local développement"
        },
        {
            "url": "https://api.eglisehub.org/api",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/admin/ministeres": {
            "get": {
                "tags": [
                    "Ministères"
                ],
                "summary": "Liste tous les ministères",
                "operationId": "f5a104859945299361c8ab0c66e72f31",
                "parameters": [
                    {
                        "name": "statut",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "actif",
                                "inactif",
                                "suspendu"
                            ]
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "eglise",
                                "ministere",
                                "organisation",
                                "para_ecclesial",
                                "mission"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Paginated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé — Super Admin uniquement"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ministères"
                ],
                "summary": "Créer un ministère",
                "operationId": "006d1e200d44e9a798c558d33fce9375",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "type",
                                    "sous_domaine"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Église Victoire"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "eglise",
                                            "ministere",
                                            "organisation",
                                            "para_ecclesial",
                                            "mission"
                                        ]
                                    },
                                    "sous_domaine": {
                                        "type": "string",
                                        "example": "eglise-victoire"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "couleur_primaire": {
                                        "type": "string",
                                        "example": "#1E3A8A"
                                    },
                                    "couleur_secondaire": {
                                        "type": "string",
                                        "example": "#FFFFFF"
                                    },
                                    "email_contact": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "pays": {
                                        "type": "string",
                                        "example": "République centrafricaine"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ministère créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Ministere"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation échouée"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/ministeres/{id}": {
            "get": {
                "tags": [
                    "Ministères"
                ],
                "summary": "Détail d'un ministère",
                "operationId": "3026abfe5168574fc7d526ff19aac471",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ministère trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Ministere"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ministères"
                ],
                "summary": "Modifier un ministère",
                "operationId": "73697c785a9e864c9a20572e9a6aa720",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "inactif",
                                            "suspendu"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ministère modifié"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ministères"
                ],
                "summary": "Désactiver un ministère",
                "operationId": "2e7f64aa1eb784593f753603ef66a98c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ministère désactivé"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/ministeres/{id}/toggle": {
            "patch": {
                "tags": [
                    "Ministères"
                ],
                "summary": "Activer ou désactiver un ministère",
                "operationId": "48b88a6b757c46ee0afa7c77992deef9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut modifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ministère actif."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Ministere"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/ministeres/{id}/stats": {
            "get": {
                "tags": [
                    "Ministères"
                ],
                "summary": "Statistiques d'un ministère",
                "operationId": "e64657e98558921a853ba25054a1d610",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stats retournées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/public/ministere": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Récupérer un ministère par sous-domaine",
                "description": "Route publique — aucune authentification requise",
                "operationId": "e1ead6e2a8d447b9f3249f07c93d2bd5",
                "parameters": [
                    {
                        "name": "subdomain",
                        "in": "query",
                        "description": "Sous-domaine du ministère",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "crc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ministère trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Ministere"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ministère non trouvé ou inactif"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Paginated": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 42
                            },
                            "data": {
                                "type": "array",
                                "items": {
                                    "type": "object"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}